requirements.txt
fuzzywuzzy
xmltodict
python-docx
openpyxl